% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{simula_fracur}
\alias{simula_fracur}
\title{simula_fracur data set}
\format{
A data frame with 200 rows and 9 variables:
\itemize{
  \item u:
  \item v:
  \item t:
  \item c:
  \item time: time observed until the occurrence of the outcome.
  \item event: variable that indicates the occurrence of the event of interest, event=1 if the event of interest occurs.
  \item int:
  \item x1_cure:
  \item x_c1:
  \item x_c2:
  \item cens: variable that indicates the outcome, cens=1 if the event of interest occurred (death due to kidney failure); cens=2 if the  dependent censoring occurred (patient received a transplant); cens=3 if administrative censoring.
  \item ident: Variable that indicates the clinic (cluster) of the patient.
  \item delta_t: indicator function of the event of interest.
  \item delta_c: indicator function of the dependent censoring.
}
}
\description{
A simulated data frame with 200 individuals, distributed in 40 clinics (clusters) with 5 individuals per clinic. Simulated dataset that mimic a kidney result study. In this simulated dataset it is hypothetically suggested the following causes of outcome: event of interest: death due to kidney failure; dependent censoring: if the patient received a transplant; administrative censoring: all other outcomes.
}
\keyword{datasets}
