% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cure.R
\name{plot_cure}
\alias{plot_cure}
\title{Plot the survival function}
\usage{
plot_cure(object, scenario = c("t", "c", "both"))
}
\arguments{
\item{object}{an object of the class "dcensoring".}

\item{scenario}{which defines the scenario in the graph (t: failure times, c: dependent censoring times, or both).}
}
\value{
a survival function graph of the fitted model.
}
\description{
This graph helps to visualize the survival function.
}
\details{
In order to smooth the line presented in the graph, we used the 'lowess' function. So, it can result in a non-monotonous survival function.
}
\examples{
\dontrun{
fit <- cure_dep_censoring(formula = time ~ x1_cure + x_c1 + x_c2 | x_c1 + x_c2,
                                     data = simula_cure,
                                     delta_t = simula_cure$delta_t,
                                     delta_c = simula_cure$delta_c,
                                     ident = simula_cure$ident,
                                     dist = "weibull")
plot_cure(fit, scenario = "t")
}

}
