% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_cure.R
\name{summary_cure}
\alias{summary_cure}
\title{Print the summary output}
\usage{
summary_cure(object, ...)
}
\arguments{
\item{object}{an object of the class "dcensoring".}

\item{...}{further arguments passed to or from other methods.}
}
\value{
a summary of the fitted model.
}
\description{
Print the summary output
}
\examples{
\dontrun{
fit_MEP <- cure_dep_censoring(formula = time ~ x1_cure + x_c1 | x_c1 + x_c2,
                                     data = simula_cure,
                                     delta_t = simula_cure$delta_t,
                                     delta_c = simula_cure$delta_c,
                                     ident = simula_cure$ident,
                                     dist = "mep")
summary_cure(fit_MEP)
}
}
